Handle bad requests better. Clean up trailing whitespace.
Requests for services that has an object as method argument
should be checked to see if an object actually was provided
and not just assume so to be able to give a proper response.

All trailing whitespace should be removed before committing
from now on. This to get rid of noisy diffs.